+++ src/test/java/com/ragin/bdd/cucumbertests/RunnerCucumber.java
@@ -1,26 +1,28 @@
 package com.ragin.bdd.cucumbertests;
 
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-import org.junit.runner.RunWith;
+import io.cucumber.junit.platform.engine.Constants;
+import org.junit.platform.suite.api.ConfigurationParameter;
+import org.junit.platform.suite.api.ConfigurationParameters;
+import org.junit.platform.suite.api.IncludeEngines;
+import org.junit.platform.suite.api.SelectClasspathResource;
+import org.junit.platform.suite.api.Suite;
 
-@RunWith(Cucumber.class)
-@CucumberOptions(
-        features = {
-                "classpath:features"
-        },
-        glue = {
-                "com.ragin.bdd.cucumber.glue",
-                "com.ragin.bdd.cucumber.hooks",
-                "com.ragin.bdd.cucumbertests.hooks"
-        },
-        plugin = {
-                "json:target/reports/cucumber/cucumber.json",
-                "html:target/reports/cucumber/cucumber.html"
-        },
-        publish = true,         // publish it to https://reports.cucumber.io. Decide if you wish it!
-        tags = "not @ignore"    // this flag works only with direct start from the IDE.
-                                // For gradle use "--tags" at build.gradle
-)
+@Suite
+@IncludeEngines("cucumber")
+@SelectClasspathResource("features")
+@ConfigurationParameters({
+        @ConfigurationParameter(key = Constants.FILTER_TAGS_PROPERTY_NAME, value = "not @ignore"),
+        @ConfigurationParameter(key = Constants.PLUGIN_PUBLISH_ENABLED_PROPERTY_NAME, value = "true"),
+        @ConfigurationParameter(
+                key = Constants.GLUE_PROPERTY_NAME,
+                value = "com.ragin.bdd.cucumbertests.glue"),
+        @ConfigurationParameter(
+                key = Constants.PLUGIN_PROPERTY_NAME,
+                value = "json:target/reports/cucumber/cucumber.json,html:target/reports/cucumber/cucumber.html"
+        )
+})
+// Its also possible to use jUnits tags instead of ConfigurationParameter
+// @ExcludeTags("ignore")
+// @IncludeTags("all")
 public class RunnerCucumber {
 }